vxlan vni should not be type uint16

+++ pkg/backend/vxlan/vxlan_network.go
@@ -88,7 +88,7 @@ func (nw *network) MTU() int {
 }
 
 type vxlanLeaseAttrs struct {
-	VNI     uint16
+	VNI     uint32
 	VtepMAC hardwareAddr
 }
